Refactored survey questions into a new component
